requirements.txt
pyyaml>=6.0
ultralytics>=8.0.0
pyzmq>=25.0.0
opencv-python>=4.8.0
torch>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pynvml>=11.0.0
prometheus-client>=0.17.0
jsonschema>=4.19.0
numpy>=1.24.0
